basic motor implementation hack
