+++ src/feagi_connector_core/data_pipeline/mod.rs
@@ -5,3 +5,4 @@ mod io_cache;
 pub use pipeline_stage::PyPipelineStage as PyPipelineStage;
 pub use pipeline_stage::extract_pipeline_stage_from_py as extract_pipeline_stage_from_py;
 pub use pipeline_stage::wrap_pipeline_stage_for_py as wrap_pipeline_stage_for_py;
+pub use io_cache::PyIOCache;
