+++ src/feagi_connector_core/mod.rs
@@ -1,2 +1,1 @@
-pub mod sensory;
 pub mod data_pipeline;
