--- src/feagi_connector_core/sensory/mod.rs
@@ -1,3 +1,0 @@
-mod sensory_cache;
-
-pub use sensory_cache::PySensorCache;
